Add support and workarounds for Capstone 5.
Capstone 5 has renamed some of the registers, which makes them less
useful for the RISC OS world - R13 and R14 are generally named SP
and LR. We now have these renamed as part of a configuration.

Capstone 5's LDR/STR handling (specifically 5.0.1) has a problem with
the returned operands; only 1 is returned. This breaks our
interpretation of the registers. We have a workaround to not try
interpreting the LDR/STR.