Restore the 'simple_disassemble.py'; allow broken down disassembly.
The 'simple_disassemble.py' example code is a good small example of
how to use the interface. Rather than just complicating things, we
should keep that tiny example and provide a more functional example
that can be used to provide an interface more like *DumpI.

The disassemble module's Disassemble object has been restructured
a little to provide two routines - the old disassemble_instruction()
has been updated to be disassemble() (returning the consumed size and
a string), and the disassemble_instruction() now returns broken down
components of the string, so that they can be reformatted if necessary.