Add the colour manipulation from Pyromaniac.
The Pyromaniac colour manipulation has been used for a few versions
of the system, and seems to work reasonably well. Extending it
to be able to generate ANSI colouring isn't too hard, and produces
a relatively good disassembly. It still needs a black background,
but only crazy people use white terminals, right ? :-)
(and you can always modify it if you prefer that)