Add a very simple setup.py to package up for use by others.
The setup.py provides a package that should let us upload it to
PyPI at some point. The name of the tool I've selected is following
the scheme of `riscos-<risc os command name>`, and as I've just
created `*DumpI` as an companion to `*Dump` in the same way that
`*MemoryI` is a companion to `*Memory`, this works.