Move example utilities to `examples`; add docs for git textconv.
The examples have been moved to the `examples` directory - and a
new example taken from RISC OS Pyromaniac has been added, just
to give a little more variety.

The README now includes information on how to configure the text
conversion so that you can see how changes have been made to binary
files. Whilst it's not especially useful to check in binaries most
of the time, this gives the ability to see how the files changed.