Initial support for the DebuggerPlus flags post-processor.
The DebuggerPlus post-processor was added to RISC OS Pyromaniac so
we might as well include it here as well. It's a relatively simple
transformation and the code is structured to be reusable, so it
can work here.

The changes here are incomplete, as we don't report help and the
documentation hasn't been updated. But it's functional.