Fix for ADR instructions not reporting right address with shifts.
When literal values were used with an ADR instruction, the wrong
address was being reported. This was due to the rotating shift not
being taken account of. This has now been fixed and we should have
the correct addresses given.

Similarly values in the boolean and arithmetic operations might not
be given the correct value in the comment. These have now been
fixed.