Added support for guessing the architecture of the file.
We can now guess the architecture of the file we are disassembling
and will usually get this right for the ARM 64 absolutes, utilities
and modules. If the format changes in the future, we will have to
update the guess_architecture function.

New examples added for the ARM32 module, ARM64 absolute and ARM64
utility.