Reworked the code to allow the decoding and access to be more flexible.
The code that provides access to the system state has been moved out
of the disassembly objects. These are now initialised independant of
the rest of the class structures. This makes things a lot easier to
work with if the host system has different access, or there are things
that are more experimental to play with.