Fix for MSR exception when spsr immediate user.
A capstone bug means that we don't return the same number of operands
for the MSR instruction when the spsr immediate is used. This meant
that we generated an exception. The instruction isn't useful in
general, so this isn't an especially important change, but it
prevents that disassembly crashing.

The solution here is just to read the first operand instead.