Improve performance by not importing capstone repeatedly.
The capstone library was being repeatedly imported each time we
went through the disassembly code. We were also populating some
dictionaries each time too. This was clearly stupid.

We now avoid this import problem and just import once.

The performance improvement under the profiler is from 32 seconds
to 6.8 seconds when disassembling the Wimp.